+++ app/components/Results/TypeResults.tsx
@@ -1,11 +1,13 @@
 import { useSearchProvider } from "@/app/provider/SearchProvider";
-import { AlbumType, ArtistType, TrackType } from "@/app/types";
-import { Box, Grid } from "@mui/material";
+import { AlbumType, ArtistType, PlaylistType, TrackType } from "@/app/types";
+import { Box, Chip, Container, Grid } from "@mui/material";
 import AlbumCard from "../Cards/Album";
 import { AlbumsLoader } from "../Skeletons/AlbumsLoader";
 import Artist from "../Cards/Artist";
 import Track from "../Cards/Track";
 import Pager from "./Pager";
+import Playlist from "../Cards/Playlist";
+import { SearchOff } from "@mui/icons-material";
 
 type TidalContentType = "albums" | "artists" | "tracks";
 
@@ -24,10 +26,14 @@ export default function TypeResults(props: TabContentProps) {
 
   return (
     <Grid container spacing={2}>
-      {data?.items?.length > 0
-        ? data?.items
-            ?.slice(0, props.limit || data?.items?.length)
-            .map((data: AlbumType | ArtistType | TrackType, index: number) => (
+      {data?.items?.length > 0 ? (
+        data?.items
+          ?.slice(0, props.limit || data?.items?.length)
+          .map(
+            (
+              data: AlbumType | ArtistType | TrackType | PlaylistType,
+              index: number
+            ) => (
               <Grid
                 item
                 xs={12}
@@ -54,12 +60,21 @@ export default function TypeResults(props: TabContentProps) {
                   />
                 ) : props.type === "tracks" ? (
                   <Track track={data as TrackType} />
+                ) : props.type === "playlists" ? (
+                  <Playlist playlist={data as PlaylistType} />
                 ) : null}
               </Grid>
-            ))
-        : !loading
-        ? "No result."
-        : null}
+            )
+          )
+      ) : !loading ? (
+        <Container maxWidth="lg" sx={{ textAlign: "center", marginTop: 3 }}>
+          <Chip
+            icon={<SearchOff />}
+            label="No result found :'("
+            sx={{ fontWeight: "bold", padding: 1 }}
+          />
+        </Container>
+      ) : null}
       {loading && (
         <Box marginTop={2}>
           <AlbumsLoader />